Move content declarations into a separate file.
